# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Japanese Kana Quiz API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuizResultInput {
  name: string;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  email?: string | null;
  /** hiragana | katakana | mixed */
  mode: string;
  /** multiple-choice | typing */
  quizMode?: string;
  score: number;
  totalQuestions: number;
  accuracy: number;
  /** Time in seconds */
  timeTaken: number;
  correctAnswers: number;
  incorrectAnswers: number;
  /** JSON string of detailed answer breakdown */
  detailedAnswers: string;
}

export interface QuizResult {
  id: number;
  name: string;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  email?: string | null;
  mode: string;
  /** @nullable */
  quizMode?: string | null;
  score: number;
  totalQuestions: number;
  accuracy: number;
  timeTaken: number;
  correctAnswers: number;
  incorrectAnswers: number;
  /** @nullable */
  detailedAnswers?: string | null;
  submittedAt: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AuthToken {
  token: string;
}

export type AdminStatsMostMissedKanaItem = {
  kana: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  totalQuizzes: number;
  averageScore: number;
  highestScore: number;
  lowestScore: number;
  mostMissedKana: AdminStatsMostMissedKanaItem[];
}

export interface DeleteResponse {
  success: boolean;
}

export type GetResultsParams = {
search?: string;
mode?: string;
sortBy?: string;
sortOrder?: string;
};