# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Japanese Kana Quiz API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  username: string;
  password: string;
}