# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Japanese Kana Quiz API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsMostMissedKanaItem } from './adminStatsMostMissedKanaItem';

export interface AdminStats {
  totalUsers: number;
  totalQuizzes: number;
  averageScore: number;
  highestScore: number;
  lowestScore: number;
  mostMissedKana: AdminStatsMostMissedKanaItem[];
}