# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Japanese Kana Quiz API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizResult {
  id: number;
  name: string;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  email?: string | null;
  mode: string;
  /** @nullable */
  quizMode?: string | null;
  score: number;
  totalQuestions: number;
  accuracy: number;
  timeTaken: number;
  correctAnswers: number;
  incorrectAnswers: number;
  /** @nullable */
  detailedAnswers?: string | null;
  submittedAt: string;
}