# lib/api-zod/src/generated/types/quizResultInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Japanese Kana Quiz API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizResultInput {
  name: string;
  /** @nullable */
  discord?: string | null;
  /** @nullable */
  email?: string | null;
  /** hiragana | katakana | mixed */
  mode: string;
  /** multiple-choice | typing */
  quizMode?: string;
  score: number;
  totalQuestions: number;
  accuracy: number;
  /** Time in seconds */
  timeTaken: number;
  correctAnswers: number;
  incorrectAnswers: number;
  /** JSON string of detailed answer breakdown */
  detailedAnswers: string;
}